APPS/SQLChat/requirements.txt
streamlit
langchain
langchain_groq
mysql-connector-python